feat: move func around to listener
since it was related more the listener than some generic events
