+++ pkg/eventcollector/listener.go
@@ -236,3 +236,70 @@ func (c *EventCollector) processEventsFromEventRecvChan(pool *tunny.Pool,
 		}
 	}(event, errs)
 }
+
+type handleEventInputs struct {
+	event  *model.Event
+	errors chan<- error
+}
+
+// handleEvent is the func used for the goroutine pool that handles
+// incoming events from the watchers
+func (c *EventCollector) handleEvent(payload interface{}) interface{} {
+	start := time.Now()
+	inputs := payload.(handleEventInputs)
+	event := inputs.event
+
+	err := c.updateEventTimeFromBlockHeader(event)
+	if err != nil {
+		return errors.WithMessage(err, "error updating date for event")
+	}
+
+	errs := c.eventDataPersister.SaveEvents([]*model.Event{event})
+	if len(errs) > 0 {
+		return errors.WithMessage(errs[0], "error saving events")
+	}
+
+	if c.crawlerPubSub != nil {
+		err = c.crawlerPubSub.PublishProcessorTriggerMessage()
+		if err != nil {
+			return errors.WithMessagef(err, "error sending message for event %v to pubsub", event.Hash())
+		}
+	}
+
+	// Update last block in persistence in case of error
+	err = c.listenerPersister.UpdateLastBlockData([]*model.Event{event})
+	if err != nil {
+		return errors.WithMessage(err, "error updating last block")
+	}
+
+	// Call event triggers
+	err = c.callTriggers(event)
+	if err != nil {
+		return errors.WithMessage(err, "error calling triggers")
+	}
+
+	// We need to get past newsroom events for the newsroom contract of a newly added watcher
+	if event.EventType() == "Application" {
+		newsroomAddr := event.EventPayload()["ListingAddress"].(common.Address)
+		newsroomEvents, err := c.FilterAddedNewsroomContract(newsroomAddr)
+		if err != nil {
+			return errors.WithMessage(err, "error filtering new newsroom contract")
+		}
+
+		errs := c.eventDataPersister.SaveEvents(newsroomEvents)
+		if len(errs) > 0 {
+			return errors.WithMessage(errs[0], "error saving events for application")
+		}
+
+		if c.crawlerPubSub != nil {
+			err := c.crawlerPubSub.PublishNewsroomExceptionMessage(newsroomAddr.Hex())
+			if err != nil {
+				return errors.WithMessagef(err, "error sending message for event %v to pubsub", event.Hash())
+			}
+		}
+	}
+
+	log.Infof("handleEvent: %v %v handled: took %v", event.ContractName(),
+		event.EventType(), time.Since(start))
+	return nil
+}
